comente-sobre/src/br/com/bluesoft/desafio/components/factories/SessionCreator.java
package br.com.bluesoft.desafio.components.factories;

import javax.annotation.PostConstruct;
import javax.annotation.PreDestroy;

import org.hibernate.Session;
import org.hibernate.SessionFactory;

import br.com.caelum.vraptor.ioc.ApplicationScoped;
import br.com.caelum.vraptor.ioc.Component;
import br.com.caelum.vraptor.ioc.ComponentFactory;

@Component
@ApplicationScoped
public class SessionCreator implements ComponentFactory<Session> {

    private final SessionFactory factory;
    private Session session;

    public SessionCreator(SessionFactory factory) {
        this.factory = factory;
    }

    @PostConstruct
    public void openSession() {
        this.session = factory.openSession();
    }

    @PreDestroy
    public void closeSession() {
        this.session.close();
    }

    public Session getInstance() {
        return factory.openSession();
    }

}
